backend/datasupport/seguridad/serializers.py
from rest_framework.serializers import ValidationError
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer

class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        if user.is_staff:
            token['is_staff'] = user.is_staff
        token['is_staff'] = user.is_staff
        return token