backend/datasupport/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
